--- src/main/java/com/dumbbelling/capstone/domain/client/auth/dto/CustomUserDetails.java
@@ -1,52 +1,0 @@
-package com.dumbbelling.capstone.domain.client.auth.dto;
-
-import com.dumbbelling.capstone.domain.client.member.Member;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-@AllArgsConstructor
-public class CustomUserDetails implements UserDetails {
-
-    private final Member member;
-
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority("ROLE_" +member.getRole().toString()));
-    }
-
-    @Override
-    public String getPassword() {
-        return member.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return member.getPhoneNumber();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
