--- src/main/java/com/dumbbelling/capstone/domain/client/member/Gender.java
@@ -1,5 +1,0 @@
-package com.dumbbelling.capstone.domain.client.member;
-
-public enum Gender {
-    MALE,FEMALE;
-}
