--- src/main/java/com/dumbbelling/capstone/domain/client/member/Role.java
@@ -1,5 +1,0 @@
-package com.dumbbelling.capstone.domain.client.member;
-
-public enum Role {
-    USER,ADMIN
-}
