--- src/main/java/com/dumbbelling/capstone/global/common/MainController.java
@@ -1,16 +1,0 @@
-package com.dumbbelling.capstone.global.common;
-
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@ResponseBody
-class MainController {
-
-    @GetMapping("/")
-    public String mainP(){
-        return "Main";
-    }
-}
