--- src/main/java/com/dumbbelling/capstone/global/config/CorsConfig.java
@@ -1,13 +1,0 @@
-package com.dumbbelling.capstone.global.config;
-
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-public class CorsConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry corsRegistry){
-        corsRegistry.addMapping("/**")
-                .allowedOrigins("http://localhost:3000");
-    }
-}
