--- src/main/java/com/dumbbelling/capstone/global/config/SecurityConfig.java
@@ -1,92 +1,0 @@
-package com.dumbbelling.capstone.global.config;
-
-import com.dumbbelling.capstone.global.security.jwt.JWTFilter;
-import com.dumbbelling.capstone.global.security.jwt.JWTUtil;
-import com.dumbbelling.capstone.global.security.jwt.LoginFilter;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.Collections;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    //AuthenticationManager가 인자로 받을 AuthenticationConfiguraion 객체 생성자 주입
-    private final AuthenticationConfiguration authenticationConfiguration;
-    private final JWTUtil jwtUtil;
-    public SecurityConfig(AuthenticationConfiguration authenticationConfiguration,JWTUtil jwtUtil) {
-        this.authenticationConfiguration = authenticationConfiguration;
-        this.jwtUtil = jwtUtil;
-    }
-    //AuthenticationManager Bean 등록
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
-        return configuration.getAuthenticationManager();
-    }
-    @Bean //비밀번호 암호화 객체 생성
-    public BCryptPasswordEncoder bCryptPasswordEncoder(){
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception{
-        //cors 설정
-        http.cors((cors)-> cors.configurationSource(
-                new CorsConfigurationSource() {
-                    @Override
-                    public CorsConfiguration getCorsConfiguration(HttpServletRequest request) {
-                        CorsConfiguration configuration = new CorsConfiguration();
-
-                        configuration.setAllowedOrigins(Collections.singletonList("http://localhost:3000"));
-                        configuration.setAllowedMethods(Collections.singletonList("*"));
-                        configuration.setAllowCredentials(true);
-                        configuration.setAllowedHeaders(Collections.singletonList("*"));
-                        configuration.setMaxAge(3600L);
-
-                        configuration.setExposedHeaders(Collections.singletonList("Authorization"));
-
-                        return configuration;
-                    }
-                }
-        ));
-        //csrf disable
-        http.csrf(AbstractHttpConfigurer::disable);
-
-        //form 로그인 방식 disable
-        http.formLogin(AbstractHttpConfigurer::disable);
-
-        //http basic 인증방식 disable
-        http.httpBasic(AbstractHttpConfigurer::disable);
-
-        //경로별 인가작업 -> 일단 작업을 위해 모두 허용
-        /*.authorizeHttpRequests((auth) -> auth
-                        .requestMatchers("/login", "/", "/join").permitAll()
-                        .requestMatchers("/admin").hasRole("ADMIN")
-                        .anyRequest().authenticated());
-         */
-        http.authorizeHttpRequests(auth->auth.anyRequest().permitAll());
-
-        //JWTFilter 등록
-        http.addFilterAfter(new JWTFilter(jwtUtil), LoginFilter.class);
-
-        //필터 추가 LoginFilter()는 인자를 받음 (AuthenticationManager() 메소드에 authenticationConfiguration 객체를 넣어야 함)
-        http.addFilterAt(new LoginFilter(authenticationManager(authenticationConfiguration),jwtUtil), UsernamePasswordAuthenticationFilter.class);
-
-        //세션을 아예 안만들도록 설정 -> 로그인해도 서버에 세션객체 저장X
-        http.sessionManagement((session)->session
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-        return http.build();
-    }
-}
