--- src/main/java/com/dumbbelling/capstone/global/security/jwt/JWTUtil.java
@@ -1,40 +1,0 @@
-package com.dumbbelling.capstone.global.security.jwt;
-
-import com.dumbbelling.capstone.domain.client.member.Role;
-import io.jsonwebtoken.Jwts;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Component
-public class JWTUtil {
-
-    private SecretKey secretKey;
-
-    public JWTUtil(@Value("${spring.jwt.secret}")String secret) {
-        this.secretKey = new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), Jwts.SIG.HS256.key().build().getAlgorithm());
-    }
-
-    public String getUsername(String token){
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("username",String.class);
-    }
-    public String getRole(String token){
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("role",String.class);
-    }
-    public Boolean isExpired(String token){
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().getExpiration().before(new Date());
-    }
-    public String createJwt(String username, String role,Long expiredMs){
-
-        return Jwts.builder()
-                .claim("username",username)
-                .claim("role",role)
-                .issuedAt(new Date(System.currentTimeMillis()))
-                .expiration(new Date(System.currentTimeMillis()+expiredMs))
-                .signWith(secretKey)
-                .compact();
-    }
-}
