--- src/main/java/com/dumbbelling/capstone/global/security/jwt/LoginFilter.java
@@ -1,89 +1,0 @@
-package com.dumbbelling.capstone.global.security.jwt;
-
-import com.dumbbelling.capstone.domain.client.auth.dto.CustomUserDetails;
-import com.dumbbelling.capstone.domain.client.auth.dto.LoginRequest;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletInputStream;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.Map;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.MediaType;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.util.StreamUtils;
-
-@Slf4j
-public class LoginFilter extends UsernamePasswordAuthenticationFilter {
-    private final AuthenticationManager authenticationManager;
-    private final JWTUtil jwtUtil;
-
-    public LoginFilter(AuthenticationManager authenticationManager,JWTUtil jwtUtil) {
-        this.authenticationManager = authenticationManager;
-        this.jwtUtil = jwtUtil;
-    }
-
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
-        //클라이언트 요청에서 username password 추출
-        LoginRequest loginRequest = new LoginRequest();
-
-        try {
-            ObjectMapper objectMapper = new ObjectMapper();
-            ServletInputStream inputStream = request.getInputStream();
-            String messageBody = StreamUtils.copyToString(inputStream, StandardCharsets.UTF_8);
-            loginRequest  = objectMapper.readValue(messageBody, LoginRequest.class);
-
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        log.info("사용자 번호 {}",loginRequest.getUsername());
-
-        String username = loginRequest.getUsername();
-        String password = loginRequest.getPassword();
-        //인증을 위해 토큰으로 변환
-        UsernamePasswordAuthenticationToken token = new UsernamePasswordAuthenticationToken(username,password,null);
-
-        return authenticationManager.authenticate(token);
-    }
-    //로그인 성공시 실행하는 메소드 (여기서 JWT를 발급하면 됨)
-    @Override
-    protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authentication) throws IOException, ServletException {
-        CustomUserDetails customUserDetails = (CustomUserDetails)authentication.getPrincipal();
-
-        String username = customUserDetails.getUsername();
-
-        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        Iterator<? extends GrantedAuthority> iterator = authorities.iterator();
-        GrantedAuthority auth = iterator.next();
-
-        String role = auth.getAuthority();
-        String token = jwtUtil.createJwt(username,role,60*60*10L);
-        response.addHeader("Authorization", "Bearer " + token);
-        /*
-            유저정보 반환로직
-         */
-    }
-
-    //로그인 실패시 실행하는 메소드
-    @Override
-    protected void unsuccessfulAuthentication(HttpServletRequest request, HttpServletResponse response, AuthenticationException failed)
-            throws IOException {
-        //커스텀 에러로 변경 필요
-        response.setStatus(401);
-        response.setContentType("application/json");
-        response.getWriter().write("{\"message\": \"잘못된 번호 혹은 비밀번호 입니다.\"}");
-    }
-}
