+++ src/main/java/com/waitit/capstone/domain/client/auth/ReissueController.java
@@ -1,0 +1,69 @@
+package com.waitit.capstone.domain.client.auth;
+
+import com.waitit.capstone.global.security.jwt.JWTUtil;
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+@ResponseBody
+@AllArgsConstructor
+public class ReissueController {
+    private final JWTUtil jwtUtil;
+
+    @PostMapping("/reissue")
+    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response){
+
+        //get refresh token
+        String refresh = null;
+        Cookie[] cookies = request.getCookies();
+        for (Cookie cookie : cookies) {
+            if(cookie.getName().equals("refresh")){
+                refresh = cookie.getValue();
+            }
+        }
+
+        if(refresh == null){
+            //response status code
+            return new ResponseEntity<>("refresh token null", HttpStatus.BAD_REQUEST);
+        }
+
+        //expired check
+        try{
+            jwtUtil.isExpired(refresh);
+        }catch (ExpiredJwtException e){
+            return new ResponseEntity<>("refresh token expired", HttpStatus.BAD_REQUEST);
+        }
+
+        //토큰이 refresh인지 확인(발급시 페이로드에 평시)
+        String category = jwtUtil.getCategory(refresh);
+
+        if(!category.equals("refresh")){
+            //response status code
+            return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
+        }
+
+        String username = jwtUtil.getUsername(refresh);
+        String role = jwtUtil.getRole(refresh);
+
+        //make new JWT
+        String newAccess = jwtUtil.createJwt("access",username,role,600000L);
+        String newRefresh = jwtUtil.createJwt("refresh", username, role, 86400000L);
+
+
+
+        //response
+        response.setHeader("access",newAccess);
+        response.addCookie(jwtUtil.createCookie("refresh", newRefresh));
+
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+}
