+++ src/main/java/com/waitit/capstone/domain/client/member/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.waitit.capstone.domain.client.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member,Long> {
+    Boolean existsByNickname(String nickname);
+
+    Member findByPhoneNumber(String phoneNumber);
+}
