+++ src/main/java/com/waitit/capstone/domain/client/member/Role.java
@@ -1,0 +1,5 @@
+package com.waitit.capstone.domain.client.member;
+
+public enum Role {
+    USER,ADMIN
+}
