+++ src/main/resources/application.properties
@@ -1,12 +1,22 @@
 spring.application.name=capstone
-#????????
+
+# database
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/capstone_db?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
 spring.datasource.username=root
 spring.datasource.password=26323174
-#??????
+
+# hibernate
 spring.jpa.hibernate.ddl-auto=create
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 
-#??? ?
+# jwt secret key
 spring.jwt.secret=vmfhaltmskdlstkfkdgodyroqkfwkdbalroqkfwkdbalaaaaaaaaaaaaaaaabbbbb
+
+# Redis
+spring.data.redis.host=${REDIS_HOST}
+spring.data.redis.password=${REDIS_PASSWORD}
+spring.data.redis.port=${REDIS_PORT}
+
+# Config import
+spring.config.import=optional:file:.env[.properties]
