+++ src/test/java/com/waitit/capstone/CapstoneApplicationTests.java
@@ -1,0 +1,13 @@
+package com.waitit.capstone;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CapstoneApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
